[Oztechan/CCC#1422] Extract ClientController
