+++ backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/controller/client/ClientController.kt
@@ -1,0 +1,5 @@
+package com.oztechan.ccc.backend.controller.client
+
+interface ClientController {
+    fun startSyncApi()
+}
