+++ backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/module/Client.kt
@@ -1,14 +1,15 @@
 package com.oztechan.ccc.backend.module
 
 import co.touchlab.kermit.Logger
-import com.oztechan.ccc.backend.repository.api.ApiRepository
+import com.oztechan.ccc.backend.controller.client.ClientController
 import io.ktor.server.application.Application
 import org.koin.java.KoinJavaComponent.inject
 
 @Suppress("unused", "UnusedReceiverParameter")
 internal fun Application.client() {
-    val apiRepository: ApiRepository by inject(ApiRepository::class.java)
+    Logger.i { "Application client" }
 
-    Logger.i { "start sync" }
-    apiRepository.startSyncApi()
+    val clientController: ClientController by inject(ClientController::class.java)
+
+    clientController.startSyncApi()
 }
