--- backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/repository/api/ApiRepository.kt
@@ -1,5 +1,0 @@
-package com.oztechan.ccc.backend.repository.api
-
-internal interface ApiRepository {
-    fun startSyncApi()
-}
